Added Support for Blacklisting Tags
Added support to define and use a tag blacklist when downloading posts to prevent posts containing any user-blacklisted tags from being downloaded. This functions like the blacklist on e621 where the user specifies a "global" set of blacklisted tags and any posts with matching tags are excluded.

In the event that the script encounters a post containing one or more blacklisted tags, it will print a message like "Post #1234 contains one or more blacklisted tags!"